+++ azure/logging_install/src/azure_logging_install/configuration.py
@@ -1,7 +1,6 @@
 import json
 import uuid
 from dataclasses import dataclass
-from logging import getLogger
 
 from .az_cmd import AzCmd, execute
 from .constants import (
@@ -10,8 +9,7 @@ from .constants import (
     STORAGE_ACCOUNT_KEY_FULL_PERMISSIONS,
 )
 from .errors import FatalError
-
-log = getLogger("installer")
+from .logging import log
 
 
 @dataclass
