+++ azure/logging_install/src/azure_logging_install/logging.py
@@ -1,0 +1,10 @@
+from logging import getLogger
+
+log = getLogger("installer")
+
+
+def log_header(message: str):
+    """Log a formatted header message."""
+    separator = "=" * 70
+    header = "\n".join(["", separator, message, separator, ""])
+    log.info(header)
